Separate ReplacementEffectWindow impl from standard trigger window
