+++ server/game/core/event/EventWindow.ts
@@ -1,5 +1,6 @@
 import type { AbilityContext } from '../ability/AbilityContext';
 import { AbilityType } from '../Constants';
+import { ReplacementEffectWindow } from '../gameSteps/abilityWindow/ReplacementEffectWindow';
 import { TriggeredAbilityWindow } from '../gameSteps/abilityWindow/TriggeredAbilityWindow';
 import { BaseStepWithPipeline } from '../gameSteps/BaseStepWithPipeline';
 import { SimpleStep } from '../gameSteps/SimpleStep';
@@ -157,7 +158,7 @@ export class EventWindow extends BaseStepWithPipeline {
 
         // TODO EFFECTS: will need resolution for replacement effects here
         // not sure if it will need a new window class or can just reuse the existing one
-        const replacementEffectWindow = new TriggeredAbilityWindow(this.game, AbilityType.ReplacementEffect, this);
+        const replacementEffectWindow = new ReplacementEffectWindow(this.game, this);
         replacementEffectWindow.emitEvents();
         this.queueStep(replacementEffectWindow);
     }
